Update 8009-pentesting-apache-jserv-protocol-ajp.md
I'd like to propose a more elegant solution of setting up a local ajp proxy i've stumbled on as an alternative or a replacement to the existing method. 

Tested on:
 - Kali linux 2020.4
 - Apache/2.4.46 (Debian)
 - ajp13
 - Apache Tomcat 9.0.19

I've managed to set up the proxy using the new method, but did not succeed in doing this by following the existing instructions. Hope this will make a good use to others :)