Update Java RMI
The content on pentesting Java RMI was not pretty useful. It was
basically a one to one copy of a blog post that discusses only one
characteristic of Java RMI.

It was replaced by an short overview on how to pentest Java RMI. This
overview contains everything one needs to know to get started.