Changing SameSite table for comprehensiveness
The table of the SameSite cookie flag used the term `Normal`, which is taken from the Invicti website, but could be misleading . Changing it in favor of `NotSet`. Further, adding `None` to the table for comprehensiveness. Finally, adding a hint to `NotSet` in the table, because of the changing default behavior of browsers.
Also fixing a typo in a keyword.