Update file-upload.md / Replace cStringIO by io
Originally, I wrote the script for the little trick "Decompress with a different name" with the module "cStringIo". I don't remember why exactly, but when tested again at home to write about this stuff, I found that cStringIo is not present anymore on python3. It works with python2.7, but `io` is better because it works with both modules.

And, I wrote " we can reuse the previous script". That's not true haha !

I think this is better to keep consistency with your previous work, and have more compatibility with different python version.

Let me know if it's ok for you. Thanks !