Open redirect page is redirecting to example.com
The open redirect page is redirection to example.com because an HTML example of open redirection vulnerability is being interpreted as HTML inside the page.
I think that the issue is about the ``` fields. I reorganized the fields to fix the issue. However, the project wasn't run in local to check my fix.